Move json tests to evmbin
