+++ evmbin/Cargo.toml
@@ -12,6 +12,7 @@ path = "./src/main.rs"
 [dependencies]
 account-state = { path = "../ethcore/account-state" }
 common-types = { path = "../ethcore/types", features = ["test-helpers"] }
+json_tests = { path = "src/json_tests" }
 docopt = "1.0"
 env_logger = "0.5"
 ethcore = { path = "../ethcore", features = ["test-helpers", "json-tests"] }
@@ -27,6 +28,7 @@ serde_json = "1.0"
 spec = { path = "../ethcore/spec" }
 trace = { path = "../ethcore/trace" }
 vm = { path = "../ethcore/vm" }
+lazy_static = "1.4.0"
 
 [dev-dependencies]
 criterion = "0.3"
