+++ evmbin/src/json_tests/src/json/blockchain/block.rs
@@ -16,7 +16,7 @@
 
 //! Blockchain test block deserializer.
 
-use crate::{bytes::Bytes, transaction::Transaction};
+use ethjson::{bytes::Bytes, transaction::Transaction};
 use super::header::Header;
 use serde::Deserialize;
