+++ json/src/lib.rs
@@ -27,5 +27,5 @@ pub mod vm;
 pub mod transaction;
 pub mod state;
 
-#[cfg(any(test, feature = "test-helpers"))]
-pub mod test_helpers;
+//#[cfg(any(test, feature = "test-helpers"))]
+// pub mod test_helpers;
