+++ rpc/src/v1/tests/mod.rs
@@ -32,7 +32,7 @@ macro_rules! extract_chain {
 	(iter $file:expr) => {{
 		const RAW_DATA: &'static [u8] =
 			include_bytes!(concat!("../../../../ethcore/res/ethereum/tests/", $file, ".json"));
-		ethjson::test_helpers::blockchain::Test::load(RAW_DATA).unwrap().into_iter()
+			json_tests::json::blockchain::Test::load(RAW_DATA).unwrap().into_iter()
 	}};
 
 	($file:expr) => {{
